Part I done, missing itoa and mapi in part II

+++ libft_subject/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: aboulest <aboulest@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/28 13:16:56 by aboulest          #+#    #+#             */
-/*   Updated: 2022/10/29 13:09:51 by aboulest         ###   ########.fr       */
+/*   Updated: 2022/10/31 00:06:54 by aboulest         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void    *ft_memchr(const void *s, int c, size_t n)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while (i < n)
